Update source code

+++ machine-learning/src/main/java/com/machinelearning/demo/api/dto/created/ProductCreatedDTO.java
@@ -23,9 +23,9 @@ public class ProductCreatedDTO {
 
     private String image;
 
-    private Long orderId;
+    private Integer orderId;
 
     private String category;
 
-    private Long categoryId;
+    private Integer categoryId;
 }
